Move functions defined in logger.h from log.cpp to logger.cpp
